requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==1.0.0
Django==3.2.15
django-heroku==0.3.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==3.3
paytm-pg==1.1.1
paytmchecksum==1.7.0


pycryptodome==3.16.0
pytz==2022.1
razorpay==1.3.0
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
